refactor: move into utils dir
